Separate exe files from other data
This is to ensure that the exe validation that happens upon every startup, is much quicker because it's only unzipping the files it needs to. And only when it needs to!

+++ MainMenu.cs
@@ -3,6 +3,7 @@ using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
+using System.Threading;
 using System.Windows.Forms;
 using static CM0102_Starter_Kit.Helper;
 
@@ -48,47 +49,49 @@ namespace CM0102_Starter_Kit {
         }
 
         private void RefreshExeFiles() {
-            string tempZipFolder = Path.Combine(Directory.GetCurrentDirectory(), "Temp");
-            string tempZipFile = tempZipFolder + ".zip";
-            File.WriteAllBytes(tempZipFile, Properties.Resources.Game);
-            new FastZip().ExtractZip(tempZipFile, tempZipFolder, null);
+            if (ExesNeedRefreshing()) {
+                string tempZipFolder = Path.Combine(Directory.GetCurrentDirectory(), "Temp");
+                string tempZipFile = tempZipFolder + ".zip";
+                File.WriteAllBytes(tempZipFile, Properties.Resources.Exes);
+                new FastZip().ExtractZip(tempZipFile, tempZipFolder, null);
 
-            if (File.Exists(Cm0102)) {
-                File.Delete(Cm0102);
-            }
-            File.Move(Path.Combine(tempZipFolder, Cm0102Exe), Cm0102);
+                if (File.Exists(Cm0102)) {
+                    File.Delete(Cm0102);
+                }
+                File.Move(Path.Combine(tempZipFolder, Cm0102Exe), Cm0102);
 
-            if (File.Exists(Cm0102Gdi)) {
-                File.Delete(Cm0102Gdi);
-            }
-            File.Move(Path.Combine(tempZipFolder, Cm0102GdiExe), Cm0102Gdi);
+                if (File.Exists(Cm0102Gdi)) {
+                    File.Delete(Cm0102Gdi);
+                }
+                File.Move(Path.Combine(tempZipFolder, Cm0102GdiExe), Cm0102Gdi);
             
-            if (File.Exists(Cm89)) {
-                File.Delete(Cm89);
-            }
-            File.Move(Path.Combine(tempZipFolder, Cm89Exe), Cm89);
+                if (File.Exists(Cm89)) {
+                    File.Delete(Cm89);
+                }
+                File.Move(Path.Combine(tempZipFolder, Cm89Exe), Cm89);
 
-            if (File.Exists(Cm93)) {
-                File.Delete(Cm93);
-            }
-            File.Move(Path.Combine(tempZipFolder, Cm93Exe), Cm93);
+                if (File.Exists(Cm93)) {
+                    File.Delete(Cm93);
+                }
+                File.Move(Path.Combine(tempZipFolder, Cm93Exe), Cm93);
 
-            if (File.Exists(Cm3)) {
-                File.Delete(Cm3);
-            }
-            File.Move(Path.Combine(tempZipFolder, Cm3Exe), Cm3);
+                if (File.Exists(Cm3)) {
+                    File.Delete(Cm3);
+                }
+                File.Move(Path.Combine(tempZipFolder, Cm3Exe), Cm3);
 
-            if (File.Exists(CmLoader)) {
-                File.Delete(CmLoader);
-            }
-            File.Move(Path.Combine(tempZipFolder, CmLoaderExe), CmLoader);
+                if (File.Exists(CmLoader)) {
+                    File.Delete(CmLoader);
+                }
+                File.Move(Path.Combine(tempZipFolder, CmLoaderExe), CmLoader);
 
-            // Cleanup
-            if (File.Exists(Cm0102Backup)) {
-                File.Delete(Cm0102Backup);
+                // Cleanup
+                if (File.Exists(Cm0102Backup)) {
+                    File.Delete(Cm0102Backup);
+                }
+                File.Delete(tempZipFile);
+                Directory.Delete(tempZipFolder, true);
             }
-            File.Delete(tempZipFile);
-            Directory.Delete(tempZipFolder, true);
         }
 
         private void RunExternalProcess(string workingDirectory, string executableFile) {
@@ -179,10 +182,17 @@ namespace CM0102_Starter_Kit {
 
         private void MainMenu_Load(object sender, EventArgs e) {
             if (!GameFolderExists()) {
+                Directory.CreateDirectory(GameFolder);
                 string gameZipFile = GameFolder + ".zip";
                 File.WriteAllBytes(gameZipFile, Properties.Resources.Game);
                 new FastZip().ExtractZip(gameZipFile, GameFolder, null);
                 File.Delete(gameZipFile);
+
+                Thread.Sleep(2000);
+                string exesZipFile = Path.Combine(GameFolder, "Exes.zip");
+                File.WriteAllBytes(exesZipFile, Properties.Resources.Exes);
+                new FastZip().ExtractZip(exesZipFile, GameFolder, null);
+                File.Delete(exesZipFile);
 
                 if (File.Exists(DefaultChangesFile)) {
                    FileInfo[] saveGames = new DirectoryInfo(DefaultGameFolder).GetFiles("*.sav");
